+++ crates/schematic/src/validate/length.rs
@@ -1,5 +1,15 @@
 use super::{map_err, Validator};
-pub use garde::rules::length::{simple::Simple, HasSimpleLength};
+pub use garde::rules::length::{
+    bytes::{Bytes, HasBytes},
+    chars::{Chars, HasChars},
+    simple::Simple,
+    HasSimpleLength,
+};
+
+/// Validate a value is within the provided length.
+pub fn in_length<T: Simple + HasSimpleLength, D, C>(min: usize, max: usize) -> Validator<T, D, C> {
+    Box::new(move |value, _, _| T::validate_length(value, min, max).map_err(map_err))
+}
 
 /// Validate a value is at least the provided length.
 pub fn min_length<T: Simple + HasSimpleLength, D, C>(min: usize) -> Validator<T, D, C> {
@@ -11,7 +21,22 @@ pub fn max_length<T: Simple + HasSimpleLength, D, C>(max: usize) -> Validator<T,
     Box::new(move |value, _, _| T::validate_length(value, usize::MIN, max).map_err(map_err))
 }
 
-/// Validate a value is within the provided length.
-pub fn in_length<T: Simple + HasSimpleLength, D, C>(min: usize, max: usize) -> Validator<T, D, C> {
-    Box::new(move |value, _, _| T::validate_length(value, min, max).map_err(map_err))
+/// Validate a value has the minimum required number of characters.
+pub fn min_chars<T: Chars + HasChars, D, C>(min: usize) -> Validator<T, D, C> {
+    Box::new(move |value, _, _| T::validate_num_chars(value, min, usize::MAX).map_err(map_err))
+}
+
+/// Validate a value has the maximum required number of characters.
+pub fn max_chars<T: Chars + HasChars, D, C>(max: usize) -> Validator<T, D, C> {
+    Box::new(move |value, _, _| T::validate_num_chars(value, usize::MIN, max).map_err(map_err))
+}
+
+/// Validate a value has the minimum required number of bytes.
+pub fn min_bytes<T: Bytes + HasBytes, D, C>(min: usize) -> Validator<T, D, C> {
+    Box::new(move |value, _, _| T::validate_num_bytes(value, min, usize::MAX).map_err(map_err))
+}
+
+/// Validate a value has the maximum required number of bytes.
+pub fn max_bytes<T: Bytes + HasBytes, D, C>(max: usize) -> Validator<T, D, C> {
+    Box::new(move |value, _, _| T::validate_num_bytes(value, usize::MIN, max).map_err(map_err))
 }